aws_lightsail_docker/connect4_uplink/requirements.txt
numpy==1.21.6
pandas==1.3.5

tensorflow==2.9.1
keras==2.9.0

anvil-uplink==0.3.36
boto3==1.24.46

h5py==3.6.0